rewrite model

+++ code/model/tag.class.php
@@ -3,66 +3,73 @@
 !defined('IN_SITE') && exit('Access Denied');
 
 class tagmodel {
-    var $db;
-    var $base;
-
-    function tagmodel(&$base) {
-        $this->base = $base;
-        $this->db = $base->db;
+    public function __construct(&$db) {
+        $this->db = & $db;
     }
 
-    function get_by_pid($pid, $limit='') {
+    public function get_by_pid($pid, $limit='') {
         $sql = "SELECT DISTINCT name FROM `problem_tag` WHERE pid=$pid ORDER BY `time` ASC";
         !empty($limit) && $sql .=  " LIMIT 0,$limit ";
 
-        $query = $this->db->query($sql);
+        $tag_array = $this->db->fetch_all($sql);
         $taglist = array();
-        while ($tag = $this->db->fetch_array($query)) {
+        foreach ($tag_array as $tag) {
             $taglist[] = $tag['name'];
         }
         return $taglist;
     }
 
-    function list_by_name($name) {
-        $query = $this->db->query("SELECT * FROM `problem_tag` WHERE name='$name'");
+    public function list_by_name($name) {
+        $tag_array = $this->db->fetch_all("SELECT pid FROM `problem_tag` WHERE name='$name'");
         $pidlist = array();
-        while ($tag = $this->db->fetch_array($query)) {
+        foreach ($tag_array as $tag) {
             $pidlist[] = $tag['pid'];
         }
         return $pidlist;
     }
 
-    function get_list($start=0, $limit=10) {
-        $taglist = array();
-        $query = $this->db->query("SELECT count(pid) as problem,name FROM problem_tag GROUP BY name ORDER BY problem DESC LIMIT $start,$limit");
-        while ($tag = $this->db->fetch_array($query)) {
-            $taglist[] = $tag;
-        }
-        return $taglist;
+    public function get_list($start=0, $limit=10) {
+        return $this->db->fetch_all("SELECT count(pid) as problem,name FROM problem_tag GROUP BY name ORDER BY problem DESC LIMIT $start,$limit");
     }
 
-    function rownum() {
+    public function rownum() {
         $query = $this->db->query("SELECT count(name) FROM problem_tag GROUP BY name");
         return $this->db->num_rows($query);
     }
 
-    function multi_add($namelist, $pid) {
+    public function multi_add($name_list, $pid, $keep_old=true) {
         if (empty($namelist)) {
             return false;
         }
-        $this->db->query("DELETE FROM problem_tag WHERE pid=$pid");
-        $insertsql = "INSERT INTO problem_tag(`pid`,`name`,`time`) VALUES ";
-        foreach ($namelist as $name) {
-            $insertsql .= "($pid,'$name',{$this->base->time}),";
+
+        $old_name_list = array();
+        if (!$keep_old) {
+            $this->db->query("DELETE FROM problem_tag WHERE pid=$pid");
+        } else {
+            $old_name_list = $this->get_by_pid($pid);
         }
-        $this->db->query(substr($insertsql, 0, -1));
-        return true;
+
+        $insertsql = "INSERT INTO problem_tag(`pid`,`name`,`time`) VALUES ";
+        $add_num = 0;
+        foreach ($namelist as $name) {
+            if (!in_array($name, $old_name_list)) {
+                $insertsql .= "($pid,'$name'," . time() . "),";
+                $add_num += 1;
+            }
+        }
+        if ($add_num > 0) {
+            $this->db->query(substr($insertsql, 0, -1));
+        }
+        return $add_num;
     }
 
-    function remove_by_name($names) {
-        $namestr = "'" . implode("','", $names) . "'";
+    public function remove_by_name($name_list) {
+        $namestr = "'" . implode("','", $name_list) . "'";
         $this->db->query("DELETE FROM problem_tag WHERE `name` IN ($namestr)");
+        return $this->db->affected_rows();
     }
+
+    private $db;
 }
 
 ?>
